feat: make the toast pop up when poruka is successful.
Decided not to use the shadcn version because of styling. It has a background color and then the custon toast is rounded, the bg is visible. I lack understanding and the will to understand how styling works right now in the toast and shadcn version of the toast so I just import the Toast from sonner instead of components/io/sonner